Refactor return types in action execution methods from (i64, Vec<u8>) to (u32, Vec<u8>)
- Updated the return type of the `execute` method in various action implementations to return (u32, Vec<u8>) instead of (i64, Vec<u8>).
- Adjusted related test cases and utility functions to accommodate the new return type.
- Removed unnecessary gas collection logic in setup functions to streamline execution.
- Ensured consistency across the VM action and context handling by updating method signatures and implementations.

+++ basis/src/interface/action.rs
@@ -1,7 +1,7 @@
 
 
 pub trait ActExec {
-    fn execute(&self, _: &mut dyn Context) -> BRet<(i64, Vec<u8>)> { never!() }
+    fn execute(&self, _: &mut dyn Context) -> BRet<(u32, Vec<u8>)> { never!() }
 }
